removed duplicate product image interfaces
